+++ packages/vm/src/backends/singlepass.rs
@@ -43,8 +43,15 @@ pub fn backend() -> &'static str {
 
 /// Set the amount of gas units that can be used in the instance.
 pub fn set_gas_limit(instance: &mut WasmerInstance, limit: u64) {
-    let used = GAS_LIMIT.saturating_sub(limit);
-    metering::set_points_used(instance, used)
+    if limit > GAS_LIMIT {
+        panic!(
+            "Attempted to set gas limit larger than max gas limit (got: {}; maximum: {}).",
+            limit, GAS_LIMIT
+        );
+    } else {
+        let used = GAS_LIMIT - limit;
+        metering::set_points_used(instance, used);
+    }
 }
 
 /// Get how many more gas units can be used in the instance.
@@ -52,4 +59,59 @@ pub fn get_gas_left(instance: &WasmerInstance) -> u64 {
     let used = metering::get_points_used(instance);
     // when running out of gas, get_points_used can exceed GAS_LIMIT
     GAS_LIMIT.saturating_sub(used)
+}
+
+#[cfg(test)]
+mod test {
+    use super::*;
+    use wabt::wat2wasm;
+    use wasmer_runtime_core::imports;
+
+    fn instantiate(code: &[u8]) -> WasmerInstance {
+        let module = compile(code).unwrap();
+        let import_obj = imports! { "env" => {}, };
+        module.instantiate(&import_obj).unwrap()
+    }
+
+    #[test]
+    fn get_gas_left_defaults_to_constant() {
+        let wasm = wat2wasm("(module)").unwrap();
+        let instance = instantiate(&wasm);
+        let gas_left = get_gas_left(&instance);
+        assert_eq!(gas_left, GAS_LIMIT);
+    }
+
+    #[test]
+    fn set_gas_limit_works() {
+        let wasm = wat2wasm("(module)").unwrap();
+        let mut instance = instantiate(&wasm);
+
+        let limit = 3456789;
+        set_gas_limit(&mut instance, limit);
+        assert_eq!(get_gas_left(&instance), limit);
+
+        let limit = 1;
+        set_gas_limit(&mut instance, limit);
+        assert_eq!(get_gas_left(&instance), limit);
+
+        let limit = 0;
+        set_gas_limit(&mut instance, limit);
+        assert_eq!(get_gas_left(&instance), limit);
+
+        let limit = GAS_LIMIT;
+        set_gas_limit(&mut instance, limit);
+        assert_eq!(get_gas_left(&instance), limit);
+    }
+
+    #[test]
+    #[should_panic(
+        expected = "Attempted to set gas limit larger than max gas limit (got: 10000000001; maximum: 10000000000)."
+    )]
+    fn set_gas_limit_panic_for_values_too_large() {
+        let wasm = wat2wasm("(module)").unwrap();
+        let mut instance = instantiate(&wasm);
+
+        let limit = GAS_LIMIT + 1;
+        set_gas_limit(&mut instance, limit);
+    }
 }